requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.1.0
blinker==1.6.2
boto3==1.35.1
botocore==1.35.1
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
cffi==1.17.0
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==1.4.0
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj-rest-auth==4.0.1
Django==4.2.19
django-allauth==0.50.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.13.0
django-environ==0.11.2
django-filter==23.2
django-ratelimit==4.1.0
django-rest-passwordreset==1.4.0
django-storages==1.14.4
django-timezone-field==5.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
executing==2.0.1
fcm-django==2.2.1
firebase-admin==6.6.0
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.53.1
fpdf==1.7.2
frozenlist==1.3.3
future==0.18.3
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.38.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==2.14.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
h5py==3.8.0
hiredis==2.3.2
html5lib==1.1
httplib2==0.22.0
hyperlink==21.0.0
icecream==2.1.3
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
incremental==24.7.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.13.1
kombu==5.3.1
libclang==16.0.0
lxml==5.3.0
Markdown==3.4.3
MarkupSafe==2.1.3
msgpack==1.1.0
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
openpyxl==3.1.2
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==23.1
pandas==2.0.2
pdfkit==1.0.0
pillow==10.4.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydyf==0.11.0
Pygments==2.17.2
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.7.0
pyOpenSSL==25.0.0
pyparsing==3.0.9
pypdf==4.3.1
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.2
qrcode==7.4.2
redis==5.2.1
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.2.2
scipy==1.10.1
sendgrid==6.11.0
service-identity==24.2.0
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
stripe==6.7.0
svglib==1.5.1
swapper==1.4.0
termcolor==2.3.0
threadpoolctl==3.1.0
tinycss2==1.3.0
tomli==2.2.1
tqdm==4.65.0
twilio==8.7.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
ua-parser==0.18.0
uritemplate==4.1.1
uritools==4.0.3
urllib3==1.26.16
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.7
whitenoise==6.4.0
wrapt==1.15.0
xhtml2pdf==0.2.16
yarl==1.9.2
zipp==3.17.0
zope.interface==7.2
zopfli==0.2.3
django-extensions==4.1
